DB

Description

Database Operations

Documentation

connectDB :: IConnection conn => conn -> IO ()

Connect the database and create all the tables

initialiseDB :: FilePath -> IO Connection

Initialise the database with a given path name

extractResp :: Either a1 [a2] -> [a2]

Extract response list from Either Left/Right as it is returned by parseRepoResponse

addRepoMany :: IConnection conn => conn -> [Either String RepoResponse] -> IO ()

Recursively calls addRepo on a list of Either RepoResponse objects

retrieveDB :: IConnection conn => conn -> [Char] -> ([SqlValue] -> b) -> IO [b]

Generalised function to retrieve and type convert an entire table from the database. Returns a list of a given data type specified by the typeConverter parameter

addContribs :: IConnection conn => conn -> (Integer, Int) -> IO ()

Adds a single contributor tuple (as provided by callContribURL in the HTTP module) to the database

addLangList :: IConnection conn => conn -> [(Integer, String, Integer)] -> IO ()

Adds a list of language tuples to the database using addLang above

addLang :: IConnection conn => conn -> (Integer, String, Integer) -> IO ()

Adds a single language tuple (as provided by callLangURL in the HTTP module) to the database

fillTotalCount :: IConnection conn => conn -> IO ()

SQL query that aggregates across all repositories and calculates the total line count and total contributors for each language. This also populates the total count table

fillLinesPerContrib :: IConnection conn => conn -> IO ()

SQL query that calculates the average number of lines per contributor for each repository. | This query then populates the LinesPerContrib table

dbTableToJSON :: (IConnection conn, ToJSON b) => conn -> [Char] -> ([SqlValue] -> b) -> IO ()

Takes a table name and its conveter (FromSQL), ecncodes the the list then writes it out to a JSON file matching it's table name

repoFromSQL :: [SqlValue] -> RepoResponse

Type converter that allows us to extract items from the database as RepoResponse objects

contribFromSQL :: [SqlValue] -> Contributor

Type converter that allows us to extract items from the database as Contributor objects

langFromSQL :: [SqlValue] -> Language

Type converter that allows us to extract items from the database as Language objects

totalFromSQL :: [SqlValue] -> TotalCount

Type converter that allows us to extract items from the database as TotalCount objects

avgContribFromSQL :: [SqlValue] -> AvgContribLines

Type converter that allows us to extract items from the database as AvgContribLines objects

topFiveLangs :: IConnection conn => conn -> IO ()

Finds and prints the top languages from the totalCount table by a number of metrics (line count, most contributors, most average lines per contributor)

topFiveContribs :: IConnection conn => conn -> IO ()

Finds and prints the top repos by contributors from the contributorResponses table

topLinesPerContrib :: IConnection conn => conn -> IO ()

Finds and prints the top repos by average lines per contributor from the linesPerContrib table

printResults :: [Char] -> [TotalCount] -> IO ()

When printing the top five results from the totalCount at the end of program execution, this function gets the relevant fields to display from the DataTypes and prints them

printContribResults :: [Contributor] -> IO ()

Gets the relevant fields and prints them for the top 5 repo's by contributor

printAvgLPCResults :: [AvgContribLines] -> IO ()

Gets the relevant fields and prints them for the top 5 repo's by average lines per contributor

retrieveRepoBetween :: (IConnection conn, Convertible a SqlValue) => conn -> a -> a -> IO [RepoResponse]

Retrieve the repos from the DB between the requested start and end repoID. This is so we call only the newly added languages_url and contributors_url in the current run. This allows us to build up a database over time of repos, language and contributor responses, without duplicating API calls, given we can only make 5000 API calls/hour.