DataTypes

Description

Data type declarations

Documentation

data RepoResponse

A data type to handle API response for public repositories.

Constructors

RepoResponse

Fields

id :: Integer

languages_url :: String

contributors_url :: String

data Contributor

A data type for contributors with multiple constructors reflecting different needs when encoding and decoding from JSON.

Constructors

ContributorFrom

Fields

login :: String

ContributorTo

Fields

repoID :: Integer

contributors :: Integer

data Language

A data type for languages with multiple constructors reflecting different needs when encoding and decoding from JSON.

Constructors

LanguageFrom

Fields

language :: String

lineCount :: Integer

LanguageTo

Fields

langRepoID :: Integer

language :: String

lineCount :: Integer

data TotalCount

A data type for extracting our derived table of total counts

Constructors

TotalCount

Fields

totalLanguage :: String

totalLineCount :: Integer

totalContributors :: Integer

linesPerContrib :: Double

data AvgContribLines

A data type for extracting our derived table of lines per contributor

Constructors

AvgContribLines

Fields

repo :: Integer

avgLinesPerContrib :: Double