HTTP

Description

HTTP Operations

Documentation

callAPI :: Exception e => String -> IO (Either e (Response ByteString))

callAPI returns an API call as an Either e (Response ByteString) when calling a given GitHub API URL. On a successful download, we can get the JSON from the Response ByteString using getResponseBody. In the case of an HTTP exception e.g. timeout, "e" will contain the HttpException details.

callContribURL :: RepoResponse -> IO (Integer, Int)

takes a RepoResponse and calls the API on the Contributor URL and returns a tuple of repoID and contributor count

callLangURL :: RepoResponse -> IO [(Integer, String, Integer)]

takes a RepoResponse and calls the API on the languages URL and returns a list of tuples of repoID, language and line count

getManyRepos :: IConnection conn => conn -> Integer -> Integer -> IO ()

As the API only returns 100 repositories at once, getManyRepos recursively calls callAPI for multiple blocks of 100 repositories. The function takes a database to write to, and a start repository ID and an end repository ID. The API is called via the URL http://api.github.com/repositories?since= where we append a repository ID to receive the 100 repositories since that ID.

removeEitherNum :: Num p => Either a p -> p

Error handling for callContribURL as parseContribResponse returns Either

splitLangResp :: t -> Either a [Language] -> [(t, String, Integer)]

Error handling for callLangURL when parseLangResponse is called as it returns an Either[]