Parser

Description

Parsing Operations

Documentation

parseRepoResponse :: ByteString -> IO (Either String [Either String RepoResponse])

Decodes raw JSON byteString from the repositories API call into a list of RepoResponse objects.

parserRepo :: Value -> Parser (Either String RepoResponse)

Parses a single decoded JSON value into an Either RepoResponse object.

parseContribResponse :: Monad m => ByteString -> m (Either String [Either String Contributor])

Decodes raw JSON byteString from the contributor API call into a list of Contributor objects.

parserContribs :: Value -> Parser (Either String Contributor)

Parses a single decoded JSON value into an Either Contributor object.

parseLangResponse :: Monad m => ByteString -> m (Either String [Language])

Decodes raw JSON byteString from the Language API call into a list of Language objects.

parserLanguages :: Value -> Parser [Language]

As the languages JSON had a varying number of unknown fields, albeit in a structured format, our parser creates a list of Language objects for each repository.

parserContribsMany :: Value -> Parser [Either String Contributor]

Maps parserContribs to an array of JSON values and returns a list of the parsed Contributor objects.

parserRepoMany :: Value -> Parser [Either String RepoResponse]

Maps parserRepo to an array of JSON values and returns a list of the parsed RepoResponse objects.